backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
autobahn==22.7.1
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
dj-database-url==0.5.0
Django==3.2.12
django-allauth==0.52.0
django-ckeditor==6.1.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-js-asset==1.2.2
django-oauth-toolkit==1.7.0
django-tinymce==3.3.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
entrypoints==0.4
fastjsonschema==2.16.1
filelock==3.8.0
fonttools==4.37.1
frozenlist==1.3.1
google-auth==2.6.2
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter-server-proxy==3.2.2
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jwcrypto==1.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
Pygments==2.13.0
PyJWT==1.7.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Send2Trash==1.8.0
shortuuid==1.0.8
simpervisor==0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
txaio==22.2.1
typing_extensions==4.1.1
Unipath==1.1
urllib3==1.26.8
virtualenv==20.16.5
vpython==7.6.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
whitenoise==6.2.0
widgetsnbextension==4.0.3
wrapt==1.13.3
yarl==1.8.1
zipp==3.8.1
